ml/requirements.txt
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.1
numpy==1.26.4
joblib==1.3.2
boto3==1.34.34
psycopg2-binary==2.9.9
geopandas==0.14.3
fiona==1.9.6
shapely==2.0.3
requests==2.31.0
meteostat==1.6.8
imbalanced-learn==0.12.0
